third_party/blink/web_tests/http/tests/xmlhttprequest/duplicate-revalidation-reload-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Make two XHRs for the resource which is already cached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result1 is "PASS"
PASS result2 is "PASS"
PASS DONE
PASS successfullyParsed is true

TEST COMPLETE
